timescale updates
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@1592 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ playground/timescale/timeinterval.cpp
@@ -73,8 +73,8 @@ static QDateTime qwtFloorDateToStep( const QDateTime &dt,
         }
         case TimeDate::Year:
         {
-            const int y = qwtAlignValue( dt.date().year(), stepSize );
-            
+            const int y = static_cast<int>( qwtAlignValue( dt.date().year(), stepSize ) );
+
             if ( y == 0 )
             {
                 // there is no year 0 in the Julian calendar
